requirements.txt
numpy==2.1.1
opencv-python==4.10.0
pillow==10.4.0
scipy==1.14.1